+++ rails_app/app/javascript/ListSet.vue
@@ -2,13 +2,11 @@
   <ul>
     <div class="calling__list">
       <p>{{ message.calling }}</p>
-      <div v-if="health_interviews_1.length === true">
-        <p>検証</p>
+      <div v-if="health_interviews_1.length">
         <NumList
+          :health-interviews="health_interviews_1"
           :modal-flag="modalFlag"
           :miss-id="missId"
-          sort-status="calling"
-          :health-interviews-1="health_interviews_1"
           @select-status="updateStatus"
           @inform-click="closeModal"
         ></NumList>
@@ -16,13 +14,11 @@
     </div>
     <hr>
     <div class="initial__list">
-      <div v-if="health_interviews_0.length === true">
-        <p>検証</p>
+      <div v-if="health_interviews_0.length">
         <NumList
+          :health-interviews="health_interviews_0"
           :modal-flag="modalFlag"
           :miss-id="missId"
-          :sort-status="initial"
-          :health-interviews-0="health_interviews_0"
           @select-status="updateStatus"
           @close-click="closeModal"
         ></NumList>
@@ -31,13 +27,11 @@
     <hr>
     <div class="pending__list">
       <p>{{ message.pending }}</p>
-      <div v-if="health_interviews_3.length === true">
-        <p>検証</p>
+      <div v-if="health_interviews_3.length">
         <NumList
+          :health-interviews="health_interviews_3"
           :modal-flag="modalFlag"
           :miss-id="missId"
-          :sort-status="pending"
-          :health-interviews-3="health_interviews_3"
           @select-status="updateStatus"
           @close-click="closeModal"
         ></NumList>
@@ -68,15 +62,12 @@ export default {
 
   data:() => {
     return {
-      // transmissionDate: {
-        // hospital_id: this.hospital_id,
-        modalFlag: false,
-        missId: "",
-      // },
-      sortStatus: "",
       health_interviews_1: [],
       health_interviews_0: [],
       health_interviews_3: [],
+       // url: `/hospitals/${this.hospital_id}/health_interviews`,
+      modalFlag: false,
+      missId: "",
       message: {
         calling: "お呼び出し中",
         pending: "保留中"
@@ -86,50 +77,24 @@ export default {
     }
   },
 
-  // created() {
-  // },
-
-  mounted() {
+  created() {
     this.fetchArray(this.hospital_id)
   },
-
-  // computed: {
-  // },
 
   methods: {
     async fetchArray(hospital_id) {
       const res_index = await axios.get(`/hospitals/${hospital_id}/health_interviews.json`)
         .then((res) => {
-          debugger
           this.health_interviews_1 = res.data.health_interviews_1
           this.health_interviews_0 = res.data.health_interviews_0
           this.health_interviews_3 = res.data.health_interviews_3
+          debugger
         })
     },
 
-    updateStatus(health_interview, e) {
+    updateStatus() {
       debugger
-      // const health_interview = value.health_interview
-      // const e = value.e
-      health_interview.guide_status.status = e.target.value
-      const hospital_id = health_interview.hospital_id
-      const id = health_interview.id
-      const url = `/hospitals/${hospital_id}/health_interviews/${id}`
-      axios.defaults.headers.common['X-CSRF-Token'] = csrfToken()
-      axios.patch(url, health_interview)
-        .then(res => {
-          this.fetchArray()
-        })
-        .catch((err) => {
-          this.selectStatus = ""
-          const id = JSON.parse(err.config.data).id
-          this.openModal(id)
-        })
-    },
-
-    openModal(id) {
-      this.modalFlag = true
-      this.missId = id
+      this.fetchArray()
     },
 
     closeModal() {
