+++ rails_app/app/javascript/components/NumList.vue
@@ -1,6 +1,6 @@
 <template>
   <div>
-    <li v-for="health_interview in healthInterviewsList" :key="health_interview.id">
+    <li v-for="health_interview in classificationHealthInterviews" :key="health_interview.id">
       {{ health_interview.guide_status.id }}
       <!-- v-model.lazy="select-status"  -->
       <select @change="selectStatus(health_interview, $event)">
@@ -21,8 +21,8 @@
 </template>
 
 <script>
-// import axios from 'axios'
-// import { csrfToken } from 'rails-ujs'
+import axios from 'axios'
+import { csrfToken } from 'rails-ujs'
 import ErrorModal from './ErrorModal.vue'
 
 export default {
@@ -32,8 +32,9 @@ export default {
   },
 
   props: {
-    // hospitalId: {
-    //   // type: ,
+    healthInterviews: Array,
+    // basisUrl: {
+    //   type: String,
     //   required: true
     // },
     modalFlag: {
@@ -44,41 +45,32 @@ export default {
       // type: ,
       required: true
     },
-    sortStatus: {
-      type: String,
-      required: true
-    },
-    healthInterviews1: Array,
-    healthInterviews0: Array,
-    healthInterviews3: Array
   },
 
   data:() => {
     return {
-      // hospital_id: this.hospitalId,
-      modalFlag: this.modalFlag,
-      missId: this.missId,
-      sortStatus: this.sortStatus,
-      // health_interviews_1: this.healthInterviews1,
-      // health_interviews_0: this.healthInterviews0,
-      // health_interviews_3: this.healthInterviews3,
+      // basisUrl: this.basisUrl,
+      // modalFlag: this.modalFlag,
+      // missId: this.missId,
+      // health_interviews: this.healthInterviews,
+      classificationHealthInterviews: [],
       // statuses,
     }
   },
 
-  mounted() {
-      debugger
-  //   const sort_status = this.sort_status
-  //   this.fetchContents(hospital_id, sort_status)
-  },
-
-  computed: {
-    healthInterviewsList: function() {
-      debugger
-      if(sortStatus === "calling") return this.healthInterviews1
-      if(sortStatus === "initial") return this.healthInterviews0
-    return this.healthInterviews3
+  watch: {
+    health_interviews: {
+      handler() {
+        debugger
+        return this.classificationHealthInterviews = this.healthInterviews
+        return
+      },
+      immediate: true
     }
+  },
+  //   this.fetchContents(hospital_id, sort_status)
+
+  // computed: {
     // params() {
     //   return {
     //     health_interview: {
@@ -89,7 +81,7 @@ export default {
     //     }
     //   }
     // }
-  },
+  // },
 
   methods: {
     // async fetchContents(hospital_id, sort_status) {
@@ -100,8 +92,28 @@ export default {
     // },
 
     selectStatus(health_interview, e) {
-      this.$emit('select-status', { health_interview, e })
+      health_interview.guide_status.status = e.target.value
+      const hospital_id = health_interview.hospital_id
+      const id = health_interview.id
+      const url = `/hospitals/${hospital_id}/health_interviews/${id}`
+      axios.defaults.headers.common['X-CSRF-Token'] = csrfToken()
+      axios.patch(url, health_interview)
+        .then(res => {
+          this.$emit('select-status')
+        })
+        .catch((err) => {
+          this.selectStatus = ""
+          const id = JSON.parse(err.config.data).id
+          this.openModal(id)
+          // this.$emit('select-status-err')
+        })
     },
+
+     openModal(id) {
+      this.modalFlag = true
+      this.missId = id
+    },
+
 
     informClick() {
       this.$emit('inform-click')
